+++ Report_App_WASM/Shared/SerializedParameters/ScheduledTaskParameters.cs
@@ -11,5 +11,5 @@ public class ScheduledTaskParameters
     public bool UseAnExcelTemplate { get; set; } = false;
     public string? ExcelTemplatePath { get; set; }
     public string? ExcelFileName { get; set; }
-    public int DataTransferId { get; set; }
+    public long DataTransferId { get; set; }
 }
